seed is a float, updated composer, trying to not bother with phpnoxdebug function in travis when running phpstan in abstract test

+++ src/Entity/Testing/TestEntityGenerator.php
@@ -79,7 +79,7 @@ class TestEntityGenerator
      */
     public function __construct(
         EntityManager $entityManager,
-        ?int $seed,
+        ?float $seed,
         array $fakerDataProviderClasses,
         \ReflectionClass $testedEntityReflectionClass
     ) {
